[SyncInit] Apache and testfield venv

+++ RB/Sync_Init/File/runtests.sh
@@ -93,7 +93,7 @@ trap end_of_script EXIT
 
 if [[ ! -d testfield ]]; then
     git clone https://github.com/Unifield/testfield.git
-    if [[ -d ~/unifield-venv ]]; then
+    if [[ -d ~/unifield-venv && ! -L ~/unifield-venv ]]; then
         pip install -r testfield/requirements.txt
     fi
     #git clone https://github.com/jftempo/testfield.git
